% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.S.PyRAD.Alleles.R
\name{Calculate.S.PyRAD.Alleles}
\alias{Calculate.S.PyRAD.Alleles}
\title{Calculate.S.PyRAD.Alleles.R}
\usage{
Calculate.S.PyRAD.Alleles(alleles.data)
}
\arguments{
\item{alleles.data}{list for each locus that is returned from the function 'read.alleles'}

\item{popID}{String to identify the particular population used to compute S statistics}
}
\description{
The function returns two lists: (1) the number of segregating sites (2) estimate of S across all loci (See pg. 97 in Wakeley)
}
\examples{
library(AlnStats)
library(stringr)

EXAMPLE
aln.data <- read.alleles(alleles.file = '~/Desktop/EXAMPLE.alleles')
}
\keyword{genetics,}
\keyword{phylogenetics,}
\keyword{population}
\keyword{pyRAD}
