% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.TajimasD.PyRAD.Alleles.R
\name{Calculate.TajimasD.PyRAD.Alleles}
\alias{Calculate.TajimasD.PyRAD.Alleles}
\title{Calculate.TajimasD.PyRAD.Alleles.R}
\usage{
Calculate.TajimasD.PyRAD.Alleles(alleles.data)
}
\arguments{
\item{alleles.data}{list for each locus that is returned from the function 'read.alleles'}
}
\description{
The function returns a list of D test for each locus provided
}
\examples{
library(AlnStats)
library(stringr)
library(pegas)

EXAMPLE
aln.data <- read.alleles(alleles.file = '~/Desktop/EXAMPLE.alleles')
}
\keyword{genetics,}
\keyword{phylogenetics,}
\keyword{population}
\keyword{pyRAD}
