% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Directory.PhylipFiles.R
\name{Read.Directory.PhylipFiles}
\alias{Read.Directory.PhylipFiles}
\title{Read.Directory.PhylipFiles.R Function to read the alignments from a directory of phylip alignments}
\usage{
Read.Directory.PhylipFiles(phy.dir, popID)
}
\arguments{
\item{phy.dir}{directory containing phylip formats}

\item{popID}{vector containing all of the samples ID that are extracted from the alignments}
}
\description{
The function returns a list of alignments obtained from a directory that contains phylip formatted alignments
}
\examples{
library(AlnStats)
library(stringr)

EXAMPLE
wanted <- c(">CV0031_0", ">CV0031_1", ">CV0310_0",">CV0310_1")
aln.data <- read.alleles2(alleles.file = '~/Desktop/EXAMPLE.alleles', popID = wanted)
S.data <- Calc.S(alleles.data = aln.data)
hist(S.data$theta.hat.vec)
hist(S.data$S.vec)
}
\keyword{genetics,}
\keyword{phylogenetics,}
\keyword{population}
\keyword{pyRAD}
