% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.PyRAD.Alleles.2.R
\name{Read.PyRAD.Alleles.2}
\alias{Read.PyRAD.Alleles.2}
\title{Read.PyRAD.Alleles.2.R Function to read the alignments of a *.alleles file generated by pyRAD (uses all individuals across alignments)}
\usage{
Read.PyRAD.Alleles.2(alleles.file, popID)
}
\arguments{
\item{alleles.file}{*.alleles file that is generated by pyRAD}

\item{popID}{vector containing all of the samples ID that are extracted from the alignments}
}
\description{
The function returns a list of alignments obtained from a *.alleles file generated by pyRAD
}
\examples{
library(AlnStats)
library(stringr)

EXAMPLE
wanted <- c(">CV0031_0", ">CV0031_1", ">CV0310_0",">CV0310_1")
aln.data <- read.alleles2(alleles.file = '~/Desktop/EXAMPLE.alleles', popID = wanted)
S.data <- Calc.S(alleles.data = aln.data)
hist(S.data$theta.hat.vec)
hist(S.data$S.vec)
}
\keyword{genetics,}
\keyword{phylogenetics,}
\keyword{population}
\keyword{pyRAD}
